Import Magento Release 1.5.0.0-beta2

+++ lib/Varien/Cache/Backend/Database.php
@@ -61,6 +61,7 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
         'adapter_callback'  => '',
         'data_table'        => '',
         'tags_table'        => '',
+        'store_data'        => true,
     );
 
     protected $_adapter = null;
@@ -136,14 +137,18 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     public function load($id, $doNotTestCacheValidity = false)
     {
-        $select = $this->_getAdapter()->select()
-            ->from($this->_getDataTable(), 'data')
-            ->where('id=:cache_id');
+        if ($this->_options['store_data']) {
+            $select = $this->_getAdapter()->select()
+                ->from($this->_getDataTable(), 'data')
+                ->where('id=:cache_id');
 
-        if (!$doNotTestCacheValidity) {
-            $select->where('expire_time=0 OR expire_time>?', time());
+            if (!$doNotTestCacheValidity) {
+                $select->where('expire_time=0 OR expire_time>?', time());
+            }
+            return $this->_getAdapter()->fetchOne($select, array('cache_id'=>$id));
+        } else {
+            return false;
         }
-        return $this->_getAdapter()->fetchOne($select, array('cache_id'=>$id));
     }
 
     /**
@@ -154,11 +159,15 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     public function test($id)
     {
-        $select = $this->_getAdapter()->select()
-            ->from($this->_getDataTable(), 'update_time')
-            ->where('id=:cache_id')
-            ->where('expire_time=0 OR expire_time>?', time());
-        return $this->_getAdapter()->fetchOne($select, array('cache_id'=>$id));
+        if ($this->_options['store_data']) {
+            $select = $this->_getAdapter()->select()
+                ->from($this->_getDataTable(), 'update_time')
+                ->where('id=:cache_id')
+                ->where('expire_time=0 OR expire_time>?', time());
+            return $this->_getAdapter()->fetchOne($select, array('cache_id'=>$id));
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -175,28 +184,30 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     public function save($data, $id, $tags = array(), $specificLifetime = false)
     {
-        $adapter    = $this->_getAdapter();
-        $dataTable  = $this->_getDataTable();
+        if ($this->_options['store_data']) {
+            $adapter    = $this->_getAdapter();
+            $dataTable  = $this->_getDataTable();
 
-        $lifetime = $this->getLifetime($specificLifetime);
-        $time     = time();
-        $expire   = ($lifetime === 0 || $lifetime === null) ? 0 : $time+$lifetime;
+            $lifetime = $this->getLifetime($specificLifetime);
+            $time     = time();
+            $expire   = ($lifetime === 0 || $lifetime === null) ? 0 : $time+$lifetime;
 
-        $dataCol    = $adapter->quoteIdentifier('data');
-        $expireCol  = $adapter->quoteIdentifier('expire_time');
-        $query = "INSERT INTO {$dataTable} (
-                {$adapter->quoteIdentifier('id')},
-                {$dataCol},
-                {$adapter->quoteIdentifier('create_time')},
-                {$adapter->quoteIdentifier('update_time')},
-                {$expireCol})
-            VALUES (?, ?, ?, ?, ?) ON DUPLICATE KEY UPDATE
-                {$dataCol}=VALUES({$dataCol}),
-                {$expireCol}=VALUES({$expireCol})";
+            $dataCol    = $adapter->quoteIdentifier('data');
+            $expireCol  = $adapter->quoteIdentifier('expire_time');
+            $query = "INSERT INTO {$dataTable} (
+                    {$adapter->quoteIdentifier('id')},
+                    {$dataCol},
+                    {$adapter->quoteIdentifier('create_time')},
+                    {$adapter->quoteIdentifier('update_time')},
+                    {$expireCol})
+                VALUES (?, ?, ?, ?, ?) ON DUPLICATE KEY UPDATE
+                    {$dataCol}=VALUES({$dataCol}),
+                    {$expireCol}=VALUES({$expireCol})";
 
-        $result = $adapter->query($query, array($id, $data, $time, $time, $expire))->rowCount();
-        if (!$result) {
-            return false;
+            $result = $adapter->query($query, array($id, $data, $time, $time, $expire))->rowCount();
+            if (!$result) {
+                return false;
+            }
         }
         $tagRes = $this->_saveTags($id, $tags);
         return $tagRes;
@@ -210,9 +221,13 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     public function remove($id)
     {
-        $adapter = $this->_getAdapter();
-        $result = $adapter->delete($this->_getDataTable(), array('id=?'=>$id));
-        return $result;
+        if ($this->_options['store_data']) {
+            $adapter = $this->_getAdapter();
+            $result = $adapter->delete($this->_getDataTable(), array('id=?'=>$id));
+            return $result;
+        } else {
+            return false;
+        }
     }
 
     /**
@@ -237,14 +252,22 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
         $adapter = $this->_getAdapter();
         switch($mode) {
             case Zend_Cache::CLEANING_MODE_ALL:
-                $result = $adapter->query('TRUNCATE TABLE '.$this->_getDataTable());
+                if ($this->_options['store_data']) {
+                    $result = $adapter->query('TRUNCATE TABLE '.$this->_getDataTable());
+                } else {
+                    $result = true;
+                }
                 $result = $result && $adapter->query('TRUNCATE TABLE '.$this->_getTagsTable());
                 break;
             case Zend_Cache::CLEANING_MODE_OLD:
-                $result = $adapter->delete($this->_getDataTable(), array(
-                    'expire_time>' => 0,
-                    'expire_time<=' => time()
-                ));
+                if ($this->_options['store_data']) {
+                    $result = $adapter->delete($this->_getDataTable(), array(
+                        'expire_time>' => 0,
+                        'expire_time<=' => time()
+                    ));
+                } else {
+                    $result = true;
+                }
                 break;
             case Zend_Cache::CLEANING_MODE_MATCHING_TAG:
             case Zend_Cache::CLEANING_MODE_NOT_MATCHING_TAG:
@@ -266,9 +289,13 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     public function getIds()
     {
-        $select = $this->_getAdapter()->select()
-            ->from($this->_getDataTable(), 'id');
-        return $this->_getAdapter()->fetchCol($select);
+        if ($this->_options['store_data']) {
+            $select = $this->_getAdapter()->select()
+                ->from($this->_getDataTable(), 'id');
+            return $this->_getAdapter()->fetchCol($select);
+        } else {
+            return array();
+        }
     }
 
     /**
@@ -385,11 +412,15 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     public function touch($id, $extraLifetime)
     {
-        return $this->_getAdapter()->update(
-            $this->_getDataTable(),
-            array('expire_time'=>new Zend_Db_Expr('expire_time+'.$extraLifetime)),
-            array('id=?'=>$id, 'expire_time = 0 OR expire_time>'=>time())
-        );
+        if ($this->_options['store_data']) {
+            return $this->_getAdapter()->update(
+                $this->_getDataTable(),
+                array('expire_time'=>new Zend_Db_Expr('expire_time+'.$extraLifetime)),
+                array('id=?'=>$id, 'expire_time = 0 OR expire_time>'=>time())
+            );
+        } else {
+            return true;
+        }
     }
 
     /**
@@ -444,7 +475,7 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
         $existingTags = $adapter->fetchCol($select);
         $insertTags = array_diff($tags, $existingTags);
         if (!empty($insertTags)) {
-            $query = 'INSERT INTO ' . $tagsTable . ' (tag, cache_id) VALUES ';
+            $query = 'INSERT IGNORE INTO ' . $tagsTable . ' (tag, cache_id) VALUES ';
             $bind = array();
             $lines = array();
             foreach ($insertTags as $tag) {
@@ -468,42 +499,46 @@ class Varien_Cache_Backend_Database extends Zend_Cache_Backend implements Zend_C
      */
     protected function _cleanByTags($mode, $tags)
     {
-        $adapter = $this->_getAdapter();
-        $select = $adapter->select()
-            ->from($this->_getTagsTable(), 'cache_id');
-        switch ($mode) {
-            case Zend_Cache::CLEANING_MODE_MATCHING_TAG:
-                $select->where('tag IN (?)', $tags)
-                    ->group('cache_id')
-                    ->having('COUNT(cache_id)='.count($tags));
-            break;
-            case Zend_Cache::CLEANING_MODE_NOT_MATCHING_TAG:
-                $select->where('tag NOT IN (?)', $tags);
-            break;
-            case Zend_Cache::CLEANING_MODE_MATCHING_ANY_TAG:
-                $select->where('tag IN (?)', $tags);
-            break;
-            default:
-                Zend_Cache::throwException('Invalid mode for _cleanByTags() method');
-            break;
-        }
-
-        $result = true;
-        $ids = array();
-        $counter = 0;
-        $stmt = $adapter->query($select);
-        while ($row = $stmt->fetch()) {
-            $ids[] = $row['cache_id'];
-            $counter++;
-            if ($counter>100) {
-                $result = $result && $adapter->delete($this->_getDataTable(), array('id IN (?)' => $ids));
-                $ids = array();
-                $counter = 0;
+        if ($this->_options['store_data']) {
+            $adapter = $this->_getAdapter();
+            $select = $adapter->select()
+                ->from($this->_getTagsTable(), 'cache_id');
+            switch ($mode) {
+                case Zend_Cache::CLEANING_MODE_MATCHING_TAG:
+                    $select->where('tag IN (?)', $tags)
+                        ->group('cache_id')
+                        ->having('COUNT(cache_id)='.count($tags));
+                break;
+                case Zend_Cache::CLEANING_MODE_NOT_MATCHING_TAG:
+                    $select->where('tag NOT IN (?)', $tags);
+                break;
+                case Zend_Cache::CLEANING_MODE_MATCHING_ANY_TAG:
+                    $select->where('tag IN (?)', $tags);
+                break;
+                default:
+                    Zend_Cache::throwException('Invalid mode for _cleanByTags() method');
+                break;
             }
+
+            $result = true;
+            $ids = array();
+            $counter = 0;
+            $stmt = $adapter->query($select);
+            while ($row = $stmt->fetch()) {
+                $ids[] = $row['cache_id'];
+                $counter++;
+                if ($counter>100) {
+                    $result = $result && $adapter->delete($this->_getDataTable(), array('id IN (?)' => $ids));
+                    $ids = array();
+                    $counter = 0;
+                }
+            }
+            if (!empty($ids)) {
+                $result = $result && $adapter->delete($this->_getDataTable(), array('id IN (?)' => $ids));
+            }
+            return $result;
+        } else {
+            return true;
         }
-        if (!empty($ids)) {
-            $result = $result && $adapter->delete($this->_getDataTable(), array('id IN (?)' => $ids));
-        }
-        return $result;
     }
 }
